requirements.txt
absl-py==1.2.0
aiohttp==3.8.3
aiosignal==1.2.0
appdirs==1.4.4
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
black==21.12b0
brunette==0.2.7
cachetools==5.2.0
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.5
cvxpy==1.2.1
cvxpylayers==0.1.5
cycler==0.11.0
daqp==0.5.1
decorator==5.1.1
diffcp==1.0.19
docker-pycreds==0.4.0
drake==1.8.0
ecos==2.0.10
executing==1.1.0
flake8==5.0.4
fonttools==4.37.3
frozenlist==1.3.1
fsspec==2022.8.2
future==0.18.2
gitdb==4.0.10
GitPython==3.1.31
grpcio==1.49.1
gurobipy==9.1.2
idna==3.4
importlib-metadata==4.12.0
iniconfig==1.1.1
ipython==8.5.0
jedi==0.18.1
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mccabe==0.7.0
meshcat==0.3.2
multidict==6.0.2
mypy==0.981
mypy-extensions==0.4.3
networkx==2.8.8
numpy==1.23.5
oauthlib==3.2.1
osqp==0.6.2.post5
packaging==21.3
pandas==1.4.4
parso==0.8.3
pathspec==0.10.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==3.0.31
protobuf==3.19.5
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.10.0
pybullet==3.2.5
pycodestyle==2.9.1
pyDeprecate==0.3.0
pydot==1.4.2
pydrake==0.2.1
pyflakes==2.5.0
Pygments==2.13.0
pyngrok==5.1.0
pyparsing==3.0.9
pytest==7.1.3
python-dateutil==2.8.2
pytorch-lightning==1.3.4
pytz==2022.2.1
PyYAML==5.4.1
pyzmq==24.0.1
qdldl==0.1.5.post2
qpsolvers==3.4.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.9.1
scs==3.2.0
seaborn==0.11.2
sentry-sdk==1.22.2
setproctitle==1.3.2
Shapely==1.8.4
six==1.16.0
smmap==5.0.0
stack-data==0.5.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tomli==1.2.3
torch==1.12.0
torchmetrics==0.6.0
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
typing_extensions==4.3.0
u-msgpack-python==2.7.1
urllib3==1.26.12
wandb==0.15.2
wcwidth==0.2.5
Werkzeug==2.2.2
yarl==1.8.1
zipp==3.8.1
